Prevent SubscriptionOffset leaks and improve type consistency
- Remove SubscriptionOffset entries on unsubscribe to prevent memory leaks
- Clean up SubscriptionOffset entries when subscribe fails (non-OK response or exceptions)
- Change SubscriptionOffset.subscriptionId from int to byte for type consistency
- Update tests to use byte cast for SubscriptionOffset constructor

This ensures proper cleanup of subscription metadata in all scenarios and
eliminates type inconsistency between the storage (int) and API usage (byte).

+++ src/test/java/com/rabbitmq/stream/impl/DeliveryTest.java
@@ -131,7 +131,8 @@ public class DeliveryTest {
 
               List<Client.SubscriptionOffset> subscriptionOffsets = new ArrayList<>();
               if (chunkOffset != subscriptionOffset) {
-                subscriptionOffsets.add(new Client.SubscriptionOffset(1, subscriptionOffset));
+                subscriptionOffsets.add(
+                    new Client.SubscriptionOffset((byte) 1, subscriptionOffset));
               }
 
               AtomicLong filteredMessageCount = new AtomicLong();
